Bugfix issue #10
The number of spns in the CoreIndex.txt file exceed the maximum
integer supported by the integer type.

The integer type has been changed from kind=2 to kind=4

+++ src/PhaseRounds.f90
@@ -61,8 +61,8 @@ MODULE PhaseRounds
 
   TYPE, PUBLIC :: CoreIndex
     ! PUBLIC
-    integer(kind = 2), allocatable, dimension(:) :: StartSnp
-    integer(kind = 2), allocatable, dimension(:) :: EndSnp
+    integer(kind = 4), allocatable, dimension(:) :: StartSnp
+    integer(kind = 4), allocatable, dimension(:) :: EndSnp
     integer(kind = 2)                            :: nCores
   CONTAINS
     ! PRIVATE
